+++ tests/utils/mod.rs
@@ -3,5 +3,6 @@
 // Created: 2026-06-28
 
 mod formatting_tests;
+mod manifest_json_tests;
 mod output_writer_tests;
 mod release_metadata_tests;